update multi turn request threshold field type from `i32` to `f32`

+++ src/protocol_types/multi_turn/request.rs
@@ -13,7 +13,7 @@ pub enum MultiTurnTestType {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct MultiTurnRequest {
     /// Safety threshold parameter
-    pub threshold: i32,
+    pub threshold: f32,
     /// Maximum conversation turns to evaluate
     pub max_turns: i32,
     /// Test case categories to evaluate
